all public types get their own files
