+++ src/game.rs
@@ -1,0 +1,5 @@
+use state::State;
+
+pub struct Game {
+    pub state: State,
+}
