+++ src/m0ve.rs
@@ -1,0 +1,7 @@
+use game::Game;
+use pos::Pos;
+
+pub struct Move {
+    pub index: (Pos, Pos),
+    pub next: Game,
+}
