+++ src/piece.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum Piece {
+    Pawn,
+    King,
+}
