+++ src/player.rs
@@ -1,0 +1,16 @@
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum Player {
+    White,
+    Black,
+}
+
+use Player::*;
+
+impl Player {
+    pub fn other(self) -> Player {
+        match self {
+            White => Black,
+            Black => White,
+        }
+    }
+}
