+++ src/pos.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Pos {
+    pub rank: u8,
+    pub file: u8,
+}
