+++ src/square.rs
@@ -1,0 +1,4 @@
+use piece::Piece;
+use player::Player;
+
+pub type Square = Option<(Player, Piece)>;
