+++ src/state.rs
@@ -1,0 +1,8 @@
+use board::Board;
+use player::Player;
+
+#[derive(Debug, Clone, Copy)]
+pub struct State {
+    pub board: Board,
+    pub player: Player,
+}
